requirements.txt
asgiref==3.7.2
certifi==2024.2.2
charset-normalizer==3.3.2
Django==4.2.3
django-jazzmin==2.6.0
django-parler==2.3
django-rosetta==0.10.0
idna==3.6
numpy==1.26.4
pillow==10.2.0
polib==1.2.0
psycopg2-binary==2.9.9
python-dotenv==1.0.1
PyYAML==6.0.1
requests==2.31.0
rosetta==0.3
scipy==1.12.0
sqlparse==0.4.4
typing_extensions==4.9.0
urllib3==2.2.1
